Fix bug in checkpoint_chunked_scan.
The bug caused the function to fail to append "remainder" steps when
the input value xs was None. This was because the check for whether or
not there was a remainder/tail chunk to concatenate checked whether
the corresponding slice of the input was None. This happened both in
the case where there is not remainder, but also in the case where the
value xs was None. Add a test that checks for this case.